+++ fed_learn/__init__.py
@@ -1,5 +1,5 @@
 from .args_helper import get_args, save_args_as_json
-from .data_sampling import iid_data_indices, non_iid_data_indices, DataHandler, CifarProcessor, BaseDataProcessor
+from .data_utils import iid_data_indices, non_iid_data_indices, DataHandler, CifarProcessor, BaseDataProcessor
 from .fed_client import Client
 from .fed_server import Server
 from .models import create_model, set_model_weights
